Refactor options pages, fix itinerary components
