--- client/components/itinerary-dates-container/index.js
@@ -1,1 +1,0 @@
-export { default } from './ItineraryDatesContainer';
