--- client/components/itinerary-times-container/index.js
@@ -1,1 +1,0 @@
-export { default } from './ItineraryTimesContainer';
