--- client/components/pages/index.js
@@ -1,6 +1,0 @@
-export { default as Page1 } from './Page1';
-export { default as Page2 } from './Page2';
-export { default as Page3 } from './Page3';
-export { default as Page4 } from './Page4';
-export { default as Page5 } from './Page5';
-export { default as Page6 } from './Page6';
