+++ client/components/single-day-itinerary/index.js
@@ -1,0 +1,1 @@
+export { default } from './SingleDayItinerary';
