+++ client/components/user-options-pages/destination-page.jsx
@@ -1,49 +1,47 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { Link, useNavigate } from 'react-router-dom';
 
-import { updateBudget } from '../../reducers/tripReducer';
+import { updateDestination } from '../../reducers/tripReducer';
 
-const Page4 = () => {
+const DestinationPage = () => {
+  // page 1
+
   const navigate = useNavigate();
 
-  const { budget } = useSelector(state => state.trip);
-  
+  const { destination } = useSelector(state => state.trip);
   const dispatch = useDispatch();
 
   const handleInputChange = e => {
     const { value } = e.target;
-    dispatch(updateBudget(value));
+    dispatch(updateDestination(value));
   }
 
   const handleKeyDown = (event) => {
     if (event.key === 'Enter') {
       event.preventDefault();
-      navigate('/form/page5');
+      navigate('/form/page2');
     }
   };
 
   return (
     <div className="bg-gray-300 rounded border-4 border-black">
-      <label className='text-2xl' htmlFor="budget">
-        Budget:
+      <label className='text-2xl' htmlFor="destination">
+        Destination:
       </label>
       <input className='typed-input'
-        type="number"
-        name="budget"
-        value={budget}
+        type="text"
+        name="destination"
+        value={destination}
         onChange={handleInputChange}
         onKeyDown={handleKeyDown}
       />
       <div>
-        <Link to='/form/page3'>
-          <button className='m-4 underline text-blue-600' type='button'>Back</button>
-        </Link>
-        <Link to='/form/page5'>
-          <button className='m-4 underline text-blue-600' type='button'>Next</button>
+        <Link to='/form/page2'>
+          <button className='m-4 text-blue-600 underline' type='button'>Next</button>
         </Link>
       </div>
     </div>
-  );
+  )
 };
 
-export default Page4;
+export default DestinationPage;
