+++ client/components/user-options-pages/index.js
@@ -1,0 +1,6 @@
+export { default as DestinationPage } from './destination-page';
+export { default as DatesPage } from './dates-page';
+export { default as ActivitiesPage } from './activities-page';
+export { default as BudgetPage } from './budget-page';
+export { default as TravelerNumberPage } from './traveler-number-page';
+export { default as TravelerTypeSubmitPage } from './traveler-type-submit-page';
