+++ client/reducers/itineraryReducer.js
@@ -7,7 +7,9 @@ const itinerarySlice = createSlice({
   initialState,
   reducers: {
     updateItinerary(state, action) {
-      state.itinerary = action.payload;
+      // console.log('payload: ', action.payload);
+      // state.itinerary = action.payload;
+      return { ...state, itinerary: action.payload }
     },
   },
 });
