+++ server/routes/itineraryRoutes.js
@@ -7,7 +7,7 @@ router.post('/build', (req, res, next)=>{
   console.log("build route invoked");
   return next();
 },authController.protect, tripController.buildTrip, tripController.saveTrip, (req, res) => {
-  res.status(201).send(res.locals.itinerary);
+  res.status(200).json(res.locals.itinerary);
 });
 
 router.get('/retrieve', authController.protect, tripController.retrieveAll, (req, res) => {
